amazonakiko-site/requirements.txt
firebase-admin>=6.0.0
requests>=2.28.0
Pillow>=9.0.0
python-dateutil>=2.8.2 